Add a level of abstraction using Writer in DbTransferLog.saveDbTransferLogFile()
DbTransferLog.saveDbTransferLogFile() needed a filename, so the output was forced to
a file.
Adding the option to pass a Writer object allow one to output the XML to stdout, using
StreamOutputWriter(System.out).

An overload has been added for reverse-compatibility.

I also added a third argument to disable purging exported logs.